chore: polish C.I. build name so it fits in output
GitHub C.I. will cut _long_ build names. So the build with name
`Build on ubuntu-latest with target aarch64-unknown-linux-musl`
will show in the interface as
`Build on ubuntu-latest with t`. This make it harder to read
the C.I. as almost all the builds appear to have the exact same
prefix.

By switching the target and the host in the name of the build,
the name becomes
`Build for aarch64-unknown-linux-musl on ubuntu-latest`.
The prefix is then different for every build.

Not perfect but easier to navigate in the GitHub C.I. interface.

Signed-off-by: Pascal Grange <pascal@grange.nom.fr>